+++ ParallaxGenLib/include/NIFUtil.hpp
@@ -9,6 +9,8 @@
 constexpr unsigned NUM_TEXTURE_SLOTS = 9;
 
 namespace NIFUtil {
+using TextureSet = std::array<std::wstring, NUM_TEXTURE_SLOTS>;
+
 // These need to be in the order of worst shader to best shader
 enum class ShapeShader { NONE, UNKNOWN, VANILLAPARALLAX, COMPLEXMATERIAL, TRUEPBR };