+++ ParallaxGenLib/include/patchers/PatcherShader.hpp
@@ -94,10 +94,4 @@ public:
   /// @param[in] NIFShape shape to apply the shader to
   /// @param[out] NIFModified if the shape was modified
   virtual void applyShader(nifly::NiShape &NIFShape, bool &NIFModified) = 0;
-
-  /// @brief apply neutral textures according to the that don't change the appearance
-  /// @param[in] OldSlots array of the slot textures
-  /// @return new array containing the applied textures
-  virtual auto applyNeutral(const std::array<std::wstring, NUM_TEXTURE_SLOTS> &Slots)
-      -> std::array<std::wstring, NUM_TEXTURE_SLOTS> = 0;
 };