+++ ParallaxGenLib/src/patchers/PatcherVanillaParallax.cpp
@@ -174,13 +174,3 @@ void PatcherVanillaParallax::applyShader(nifly::NiShape &NIFShape, bool &NIFModi
     NIFModified = true;
   }
 }
-
-auto PatcherVanillaParallax::applyNeutral(const std::array<std::wstring, NUM_TEXTURE_SLOTS> &Slots)
-    -> std::array<std::wstring, NUM_TEXTURE_SLOTS> {
-
-  array<wstring, NUM_TEXTURE_SLOTS> NewSlots = Slots;
-
-  NewSlots[static_cast<size_t>(NIFUtil::TextureSlots::PARALLAX)] = L"textures\\parallaxgen\\neutral_p.dds";
-
-  return NewSlots;
-}