+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Mods that do not have any special shader type can be prioritized now
 - When mixing shaders with alternate texture records meshes will now be duplicated where required
 - Meshes that have higher priority than textures can be patched now
+- Plugin patching will also be considered when evaluating mods for conflicts now
 
 ## [0.7.3] - 2024-12-09